requirements.txt
# requirements.txt
websockets
kafka-python
psutil
python-json-logger
